requirements.txt
# Core Streamlit
streamlit>=1.28.0

# LangChain and related packages
langchain>=0.1.0
langchain-core>=0.1.14
langchain-community>=0.0.34
langchain-openai
openai

# Vector databases and embeddings
faiss-cpu>=1.7.4

# Document processing
pypdf>=4.0.1
docx2txt>=0.8
unstructured>=0.12.0
python-magic-bin>=0.4.14  # Better cross-platform support than python-magic

# Environment and utilities
python-dotenv>=1.0.0

# Text processing
chardet>=5.2.0  # For better encoding detection